refactor : statistics 도메인 생성 및 구조 변경
