+++ src/main/java/org/swyp/dessertbee/user/repository/UserRepository.java
@@ -5,10 +5,9 @@ import jakarta.validation.constraints.NotNull;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import org.swyp.dessertbee.user.dto.response.UserStatisticsResponseDto;
+import org.swyp.dessertbee.statistics.common.dto.user.UserStatisticsResponseDto;
 import org.swyp.dessertbee.user.entity.UserEntity;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
